event types: Apply clarifications on update_message
From zulip/zulip#20693.  It turns out that almost all the properties
on these events are optional, depending on the nature of the edit.

Most of the updates our code needed for handling this came in the
preceding few commits; particularly in the two recent commits that
edited this type, applying some of the changes corresponding to that
API-docs PR zulip/zulip#20693.  (We did that where the code changes
were complex enough that they call for tests, and for the tests to
be well-typed we need the corresponding updates to the type.)

We add one small logic fix, and then simplify some tests to reflect
properties that are no longer required.