stream [nfc]: Delete redundant, inefficient getSubscribedStreams
This just takes each of the Subscription objects we have, and
spreads the corresponding Stream object on top of them.  But the way
these objects work in the server API is that a Subscription object
already contains all the properties of the corresponding Stream in
the first place.  So the end result should have exactly the same
contents as the Subscription object we started from.

When this selector was introduced in 27483c9a8 back in 2017, the
motivation was apparently "because streams change events are updated
in streams" -- i.e., because we were *not* keeping the Subscription
objects up to date when things changed on the stream.

But in that case the real bug was that we were keeping around an
object with out-of-date data; having only an alternative wrapper
off to the side where one can get an up-to-date version is a recipe
for an ongoing stream of bugs where we use the version that's in
the obvious place, not realizing it can be stale.

Fortunately that did get corrected the next year, in 152b06e25.
Since then, the Subscription objects are just as up to date as
the Stream objects, and this selector has no reason to exist.
Drop it, and have its callers get the subscription data directly.