android notif [nfc]: Update API comment on "user"
This comment mentioned server 2.0, but just as the latest as of
when it was written.  This field was deleted at a later version,
so note that down; and the real prompt for deleting the comment
will be when "user_id" is always present and can be relied on.
So copy the "TODO(server-2.1)" from there.