NavBarBackButton: Use `useNavigation` instead of NavigationService
We want to get rid of NavigationService anyway.

Our immediate reason to do this is to fix a bug where the navigation
failed with the following message:

  ExceptionsManager.js:180 The action 'POP' with payload {"count":1}
    was not handled by any navigator.

  Is there any screen to go back to?

  This is a development-only warning and won't be shown in
    production.

Seems like the Hook ends up being better at getting the best, most
locally relevant `navigation` object for the component's location in
the React Navigation hierarchy. Not sure precisely why. It could
easily be the case if it's using React Context under the hood; I bet
it is. Anyway…

The Hook's ability to provide the appropriate `navigation` object
may be considered too much "magic" or action-at-a-distance, and Flow
can't track the type of the `navigation` object we end up getting,
to help us confirm that the "back" behavior is what we want.

These are reasonable concerns. But also, we can hope that it's
obvious to `NavBarBackButton`'s callers what the intended
`navigation` object is, and what "back" behavior we'll end up
getting from it. And, failing that, the docs claim that all
`navigation` objects at least *have* a `goBack` method:
  https://reactnavigation.org/docs/5.x/navigation-prop/
(It doesn't always do the same thing as on a stack nav; see
`backBehavior` on drawer navs and bottom-tab navs.)