mockup: Drawer nav on the leftmost bottom tab
Implemented:
- Four buttons at the top gone :)
- Hamburger icon to open the drawer from drawer nav's default screen
  (labeled "Home" in the drawer)
- Swipe from left to open the drawer
- backBehavior="initialRoute" (my "resolved concern" in
  https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/left-side.20drawer.20nav/near/1308981 )

Not implemented:
- Full app bar (just a quick hamburger menu)
- Resetting to default screen (dubbed "Home") when you go to a
  different bottom tab and come back (don't think this would be too
  hard?)
- Sane deduplication of lots of code :P
- …probably quite a bit more that I'm not remembering right now, but
  it's a quick mockup ;P

Code details for developers:

I used copy-and-paste to an extent that would be irrational for
production code. :) It's why I was able to do a mockup so quickly.
It'll be tricky but informative to deduplicate well. I can talk more
about why I think so, or you might reach the same conclusion by just
studying the code and the APIs. (Maybe `git diff` has something to
help you see what tweaks I made after copy-pasting.)

Anyway, here's a sketch of how the code got to where it is:

- HomeScreen was sitting on the main-tabs navigator, as the 'home'
  screen (the leftmost tab). Now, a new HomeDrawerNavigator is in
  its place there.
- The stuff that was being rendered on HomeScreen has moved to a new
  DefaultScreen, minus the four buttons at the top.
- I made an AllNarrowScreen, StarredNarrowScreen,
  MentionedNarrowScreen, and SearchInHomeDrawerScreen. The first
  three were copied from ChatScreen (with the narrow hard-coded);
  the last, from SearchMessagesScreen.