react nav: Do an installation step we forgot about (oops)
And add a note about a step specific to react-native-gesture-handler
that seems fine to omit for now.

See React Navigation's setup instructions at
  https://reactnavigation.org/docs/5.x/getting-started#installing-dependencies-into-a-bare-react-native-project :

  To finalize installation of `react-native-gesture-handler`, add
  the following at the **top** (make sure it's at the top and
  there's nothing else before it) of your entry file, such as
  `index.js` or `App.js`:

  ```js
  import 'react-native-gesture-handler';
  ```

  > Note: If you are building for Android or iOS, do not skip this
  > step, or your app may crash in production even if it works fine
  > in development. This is not applicable to other platforms.

Well, better late then never, I guess. I haven't noticed anything
change with this fix, or anything in `main` that doesn't work
without the fix. But it's what we should be doing, so do it. The
latest React Nav version (currently 6.x) doesn't mention this in the
overall setup instructions, but does mention it in pages for some
individual stack navigators, including
  https://reactnavigation.org/docs/stack-navigator/ .