nav: Have main-tabs screens consume top inset individually
Instead of doing it centrally in MainTabsScreen.

Most likely, for #5089 (app bar on main views), we'll ask each
screen to render an app bar component specific to the screen (if
only to have a different title, e.g.), instead of placing one app
bar centrally in MainTabsScreen. Since the app bars will want to
consume the top inset (with a contentless background), this change
is in the right direction.

We do it now to help with a drawer-nav mockup; the screens on the
drawer nav were unhelpfully stretching to consume the top inset,
when MainTabsScreen was already doing so, resulting in awkwardly
empty space.