nav: Omit "topic list", "stream info" buttons on unknown streams
If we actually make it to either of these screens (TopicListScreen or
StreamSettingsScreen) with an unknown stream, they'll try to look up
our data on the stream and will promptly crash.

We're allowing this ChatNavBar component to tolerate unknown streams,
because one can sometimes navigate to a narrow in such a stream and
we try to represent that as a proper screen: ChatScreen will show
InvalidNarrow in that case instead of MessageList, and will show a
ChatNavBar as usual.  But then let's draw the line at navigating to
any other screens.  This one already delivers the message that
there's nothing here; no need to have other screens handle this case
just to deliver the same message.

At the moment, these buttons don't do anything in this case anyway:
if the stream is unknown, then the callback won't find the stream
when it looks it up, and will silently do nothing.  But in our
migration toward stream IDs, we'll shortly have these callbacks
navigate directly by stream ID rather than have to look up the name
in our streams data.  Without this change, that would result in
navigating to the destination screen with an invalid stream.