mute: Stop taking stream name in isTopicMuted
In the previous commit, we already stopped looking at this argument
in favor of the stream ID.

This change should be NFC except in the case where the given stream
didn't exist in our data structures, at the several call sites that
were looking it up in order to find the name.  Those now simply go
and look up the mute status, where they presumably find nothing.

In particular if it's possible to reach the call site in
constructTopicActionButtons in that case, then previously we would
crash; now we offer to mute the topic.