api types [nfc]: Update version comment on user_status in initial data
This comment mentioned version 1.9.1, but only as the then-latest
release, which didn't yet support the feature.

That's a bit before the start of good changelogs in our API docs,
and the "user status" feature isn't mentioned in the API changelog.
But the overall Zulip Server changelog at version 2.0.0:
  https://zulip.readthedocs.io/en/latest/overview/changelog.html#id32
indicates the feature was new in that version.