notif types: Mark user_id optional; document status of types
This property should be optional, because the code still supports
server versions where it's absent.  (As reflected in the comment
there, and more concretely in FcmMessage.kt where it handles the
same user_id property.)

Then marking it optional raises the question: why doesn't Flow
tell us about an error?  In fact we were already not assuming the
property is present: extract.js explicitly handles the case where
it's absent.  But also it turns out that even if we were making
that assumption, Flow wouldn't be able to make the connection,
because we don't use these types in code.  That's different from
how we use most of our API types -- so document it explicitly
with some comments, to help avoid surprises in the future.