stream [nfc]: Give StreamItem callbacks a stream-like object
This way, the use sites are now in the same situation as they would
be if passed a whole Stream object.  That makes it entirely their
responsibility whether they use the stream's name, either
appropriately or inappropriately.

As it happens, there's one callback that uses it appropriately (to
generate a Zulip-flavored-Markdown stream reference in autocomplete)
and one that uses it inappropriately.  The latter is because that's
what some API methods take, and already has appropriate comments.