eslint: Remove flowtype/generic-spacing lint.
This conflicts with our prettier config, and doesn't seem to be very
useful - we frequently needed to disable it. It was added in a1edcaa1
without any explanation, but it seems to largely have been a copy of the
recommended config at the time.

See https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/fmt.2Ftest.20prettier.20mismatch/near/1326785