react-nav libdefs: Move bottom-tabs to its own .flow.js file
Inspired by 007dea350.

I took everything in the `declare module` block (thankfully there
was just one such block), moved it to the new file, and removed
`declare` everywhere, as in `declare export type …`.

And move the tests closer to the file, in `types/__flow-tests__/`.
When we started adding these test files, in ad82c4c31, we said it'd
be best for the tests to live closer to the code being tested.