RootErrorBoundary: Don't try to send possibly non-JSONable extras to Sentry
Flow v0.149 would rightly point out that parts of the inexact object
at `errorInfo` might not be JSONable, and therefore not suitable to
pass to `logging.error` in the second argument.

So, pass only the part we do expect to be JSONable, and add a
`logging.warn` to tell us about any other data in `errorInfo` we
might like to send in the `logging.error`.