stream: Memoize sorting in StreamListCard and SubscriptionsCard
These lists can be kind of long (for example, there are a few hundred
streams on chat.zulip.org).  So when we do the work of allocating new
copies of them and sorting those, it's good to memoize that work.

This memoization should be effective because the inputs here come
straight from the state, and only change when the relevant data has
actually changed.

While we're at it, also memoize the similar work of separating pinned
from unpinned streams in SubscriptionsCard.

This change should be NFC apart from any performance effect.