ios notif [nfc]: Simplify out some redundant logic
It's true that another iOS notifications library might have a
different interface in this respect.  But if someday in the future we
switch to a different iOS notifications library with a different
interface, we'll be updating this code for the new set of expectations
in any case -- there's no gain from attempting to accommodate in
today's code one of the many possible ways the interface could differ.

Moreover, as far as this particular point of the interface is
concerned… none of the subsequent code actually notices or cares
whether there's an `aps` property on the object.  It takes the `zulip`
property, and does nothing with the rest of the object.  So going and
removing an `aps` property first has no meaningful effect.