event [nfc]: Keep update_message event verbatim on action, don't spread
We're going to add another property here, to carry some precomputed
information that multiple reducers will want.  Seems cleanest to do
that separately from the event, rather than mixing more properties
into it.

For the moment this structure looks duplicative of the `type: EVENT`
case, but that will change when we add another property.