nav: Update ChatScreen narrow on topic move
This covers the most important bit: when applicable, the narrow the
ChatScreen is looking at gets updated to follow the moved messages.

Left open are a couple of other things to do, related to the
compose box.  I think I'm content to close the issue without those,
and open follow-up issues for them.  Include TODO comments for them.

Fixes: #5251