imports [nfc]: Cut the common/index.js indirection
This wasn't doing anything useful -- in particular, if you know the
name of the component you're importing, then you already know the
name of the specific module it's defined in -- and it tends to
create import cycles.  Cut it out.

Done mostly automatically, like so:

  $ perl -i -lne '
      if (m<^import\s* \{\s*
               ( (?: \w+ ,? \s*? )+ )
             \s*\}\s* from\s+ (\S*/common).;$
           >x) {
        my ($list, $path) = ($1, $2);
        for my $name (split /,\s*/, $list) {
          print "import $name from $path/$name'\'';";
        }
      } else {
        print
      }
    ' src/**/*.js

  $ perl -i -lne '
      if (m<^import\s* \{\s*
               ( (?: \w+ ,? \s*? )+ )
             \s*\}\s* from\s+ (.\.).;$
           >x) {
         my ($list, $path) = ($1, $2);
         for my $name (split /,\s*/, $list) {
           print "import $name from $path/$name'\'';";
         }
       } else {
         print
       }
     ' src/common/*.js

plus manually handling PasswordAuthScreen, and deleting the
indirection file itself.

After this change, the majority of the runtime import cycles we had
are gone.

At type-checker time (as displayed by tools/import-cycles), the giant
cycle-blob of 146 files is now broken up, with a little over half its
files in one of several smaller blobs and the rest now free of cycles.
Including the one other, 10-file, blob we had had, the overall set of
blobs is now:

  43 files around src/nav/AppNavigator.js
  25 files around src/selectors.js
  10 files around src/types.js
   6 files around src/webview/MessageList.js
   4 files around src/api/apiTypes.js
   3 files around src/boot/store.js