unread: Use stream ID for a React key, not stream name
These `key` properties just get used as section keys in the
SectionList we create in UnreadCards.  SectionList docs:
  https://reactnative.dev/docs/sectionlist#section
don't quite say that they get used as the `key` pseudo-prop of some
React components internally; but they do say that they have basically
the same effect as such React keys would.

This change is arguably not *quite* NFC, because it changes how these
keys behave if a stream is renamed.  Previously it'd make a fresh
React component -- or if one stream was renamed to another's old name
between rerenders, it'd reuse the other stream's old component.  Now
it'll reuse the same stream's old component.

I believe it still ends up being NFC -- apart from a possible perf
improvement, which is negligible because it's such an uncommon
occasion and the old behavior wasn't having any catastrophic perf
impact -- because I don't think there's any local state on the
components we have in the subtrees these act as the keys of.