UserStatusScreen: Use plain-old navigation.goBack()
We don't like using NavigationService (#4417), so this is nice to be
able to do.

Not *quite* NFC: if we somehow manage to have two consecutive
UserStatusScreens at the top of the stack, we'll now just pop one of
them instead of both, because we lose `navigateBack`'s special logic
with its `sameRoutesCount` value. But that logic is designed for
`ChatScreen` -- we do expect to have multiple `ChatScreen`s at the
top of the stack sometimes. We don't expect that with
`UserStatusScreen`s.