action-sheet [nfc]: Consolidate constructMessageActionButtons
The name "constructNonHeaderActionButtons" wasn't working -- it
worked when the only action sheets were on a recipient header or on
a message, but not so much anymore as there are several other action
sheets that would properly be covered by "non-header".

We could have some circumlocution for "Message or Outbox", but
instead just say "message" for that as we usually do.