topic [nfc]: Fix internal references to "subject"
The bulk of the places we mention "subject" are direct references to
either the property of that name on `Message` or somewhere else in
the server API.  Most of those, and in particular `Message#subject`,
are still that way in the current API.

Those will be an annoying giant conversion, so I have no desire to be
in a rush to do them.  In particular, for any given piece of API, I
won't want to make a conversion before we've at least settled on the
API change so that we can convert to what we know is the future API.

However, there are a handful of places where we say "subject" to
mean "topic" and it's entirely internal to the mobile app.  So
here we go ahead and fix those.