user-statuses model: Export `getUserStatuses` after all
But with a warning in jsdoc, so callers don't misuse it (the concern
that led us to un-export it in 1ba51386d).

The message list's BackgroundData has so far been a collection of
stuff picked from PerAccountState, not an entire PerAccountState
itself. It'll be easiest to not break that pattern when we start
showing emoji statuses in the message list, coming soon.

Discussion at
  https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/user-statuses.20model/near/1340249