ComposeBox: Put effects of message-input change in componentDidUpdate
These are properly seen as effects: things that happen as a
consequence of the message-input value changing. So
`componentDidUpdate` is a fine place for them. (Later, when this is
a function component, we'll put them in a `useEffect`.)

Also, we'll soon have a reason to inline `handleMessageChange` at
one of its callsites, in setMessageInputValue. This makes it
possible to do so without duplicating all these lines of code. (The
body of handleMessageChange just has one line after this.)

We don't want this chunk of code in componentDidUpdate to run on
every props or state update; we just want it to run when the message
state changes. So, we put it inside a
`prevState.message !== message` check. Since we don't set the
message state outside handleMessageChange, there won't be cases
where the chunk of code is triggered now but it wasn't before.
Possibly this removes some cases where it was triggering before;
those would be cases where handleMessageChange was called with a
`message` that matched the current message state. If so, that's
fine, because if such calls were indeed happening, nothing would
have come of them because sendTypingStop, sendTypingStart, and
draftUpdate are idempotent.

It's not *quite* NFC because these actions run an extremely short
time later than they did before. There should be no problem there,
because the state that can be affected (is-user-typing and the saved
draft) aren't read by this component. (Well, the draft potentially
is, but only at the very beginning of the instance's lifecycle.)