msglist tests: Use `stream1` in single-messages tests
The previous stream object was in an awkward in-between position:
its ID happened to match the `stream1` and `stream1`'s subscription
in the state (`baseState`) that the `check` function used as a
default. But the stream's name differed in an arbitrary and
unimportant way. (Both stream objects are "nonrandom" enough for
current purposes.)

So, just use `stream1`.