StoreHydratedGate [nfc]: Rename from HideIfNotHydrated; repurpose as singleton
It looks like this component was written to be flexible and
reusable, with callers giving their own PlaceholderComponent as
desired.

So far we've only used it once. And that's not a coincidence:
encapsulating this check in just one place is a sensible design; see
  https://github.com/zulip/zulip-mobile/blob/master/docs/architecture/crunchy-shell.md

Not doing so, and instead having to check `isHydrated` in multiple
places, risks being confusing.

So, let this component settle into its role as a singleton.