initial data types: Add TODOs to update to modern `user_settings` object
And mark write-only (with the contravariance sigil, `-`) the
deprecated properties that appear in our types but that we're not
currently using, so that we won't accidentally start using them.

This should be all of them. I checked all 36 results of the search
term [did not include user_settings_object] on
https://zulip.com/api/register-queue on 2022-03-15. Several of the
36 were properties that we don't use and that don't appear in our
types at all. (No use adding those, I think, just for them to be
removed when we do the migration.)

Related: #4933