example data [nfc]: Make names/emails/avatars explicit in example users
The point of adding " User" like we do is to make these names
somewhat meaningful when they come up in test-failure output.
Given that, it seems best to have the actual names, e.g.
"Other User", show up in the codebase, findable by search.

Similarly for the emails and avatar_url values.