event [nfc]: Check eventToAction for exhaustiveness on event type
Awkwardly this actually requires the cases to use string literals
directly, rather than the named constants.  I prefer named constants
for the sake of greppability... but when required to choose, checking
for errors is more important.

The fact that Flow passes after this change means that the `switch` is
covering all the event types our API code knows about.  There are a
number of others the server can send, though.  Next we'll add all of
those to the list, and explicitly call out the ones we're ignoring.