event: Add rendered_description quirk on stream updates of description
The API returns an additional property, rendered_description, when the
updated property of a stream update is description. This commit reflects
this in the StreamUpdateEvent type.

We don't currently interact with the rendered_description property of
streams at all, so there's nothing to do in the reducer.  Just note in
a comment on the Stream type so that we see it if we every start using
that property in the future.