DevAuthScreen: Get error message correctly; type-check `catch` bodies
Applying types here made it clear that this code was not actually
working!  It was looking for the error message at `err.data.msg`,
implicitly assuming it was one of our `ApiError` objects... but in
fact, `ApiError` puts the error message at the standard `message`
property (following `Error`), and doesn't duplicate it on `data`.