android build: Upgrade Android Gradle Plugin to v4.2.1
Done to follow the template-app change in
facebook/react-native@7599593b3, on the path to the RN v0.65
upgrade.

Release notes here:
  https://developer.android.com/studio/releases/gradle-plugin#4-2-0

Quite a few changes announced; hopefully they're all good ones. Our
tests pass, and a debug build on Android built and ran fine for me.

Greg points out that this one is interesting and good:

> Starting in AGP 4.2, it is no longer possible to override Gradle
> properties from subprojects. In other words, if you declare a
> property in a gradle.properties file in a subproject instead of
> the root project, it will be ignored.

For more details on that, see
  https://github.com/zulip/zulip-mobile/pull/5324#discussion_r842261767 .

The release notes say Java 8 is now specified by default. So, follow
the RN commit and remove the config lines that specify Java 8.

The RN commit also bumps the template app's Gradle version to 6.9.
We already did that for ours, in fb0cf3103.