nav: Enable react-native-screens
This greatly improves performance on Android with many screens on
the nav stack!  On iOS, the effect is less dramatic, but still
significant.

Chris investigated this back in 2020 (#4111), and sent a draft PR
to enable it (#4161), but we never followed up, oops.  Better late
than never.

Experiment to repro the perf difference:

  1. push like 30 ChatScreen screens to the stack, by following a
     loop of internal links;

  2. go to one with some backlog (by following the "up" icon to the
     stream message-list);

  3. try scrolling around.

Before, on Android: extremely laggy.  Also pretty laggy even just
following those links, after the first few.

After: pretty much as smooth with 30 screens on the stack as with the
first one.

On iOS, things seem to already be about as smooth with 30 screens as
with one, even before this change; in particular, scrolling through
the message list is very smooth.  (That's even though the app in my
iOS testing is fairly sluggish on the first few navigations.  I was
using the simulator on an underpowered Mac Mini, vs. the Android
emulator on my fast Linux desktop.)

But Xcode does report (under the "Debug View Hierarchy" button) that
before the change, tons and tons of views remain mounted when many
screens are on the stack; after the change, they don't.  I reproduce
the same observations Chris reported a couple of years ago here:
  https://github.com/zulip/zulip-mobile/issues/4111#issuecomment-643519230

More concretely, the memory usage (which Xcode reports at the same
time) gets cut way down: around 700 MB before, and around 300 MB
after.  If nothing else, that should mean that the app is less
likely to get killed by the OS if the user switches away -- so is
more likely to pick up immediately where they left off, rather than
have to reload from scratch, if the user switches back.

Fixes: #4111