react-native-webview libdef: Remove some now-dangling references
I should've caught these in 684498afd; I think I assumed these
definitions were purely exported, without anything using them within
the same file. Flow didn't catch the dangling references then
because, at this version, it doesn't properly report errors in these
libdef files.

I think the conclusion of 684498afd -- that the definitions should
be removed -- is unchanged, because of the low quality of the
definitions. They ignore the interesting imperative methods that are
specific to the webview, some of which differ per-platform.