api [nfc]: Have createStream use params objects
One for stream attributes, and one for miscellaneous options.

The create-stream API is a little awkward in that it piggy-backs on
the subscribe-to-stream API. Here's the doc:
  https://zulip.com/api/create-stream

> You can create a stream using Zulip's REST API by submitting a
> [subscribe](/api/subscribe) request with a stream name that
> doesn't yet exist and passing appropriate parameters to define the
> initial configuration of the new stream.

So, make our code a bit clearer by bundling the attributes of the
to-be-created stream into an object. We're about to add more
attributes, to sync with the API doc.

While we're at it, order the attributes by their appearance in the
doc.