api: Remove more param-defaulting within our api.createStream binding
With similar reasoning as in the previous commit, except these are
in `options` instead of `streamAttributes`.

This time, one of the properties doesn't have a default applied by
the single caller: `announce`. So this is a behavior change: we're
now sometimes omitting `announce`, which lets the server decide what
behavior to fall back on. Currently, it's documented as falling back
to `false` (and that should apply for all non-ancient servers):
  https://zulip.com/api/subscribe#parameter-announce