CreateStreamScreen: Let server apply its own fallback for `principals`
This seems like a fine place to let the server control the behavior,
and it's documented to give the same behavior anyway. See
  https://zulip.com/api/subscribe#parameter-announce :

> […] If not provided, then the requesting user/bot is subscribed.

We can start passing `principals` again when we eventually have a UI
for choosing multiple users to be subscribed, as the web app does.

For now, it's redundant and unnecessary for us to request the same
behavior the server would give anyway. Especially since we're asking
for it in a way that's discouraged: we use an email instead of a
user ID.