emoji: Switch to shared code for finding matches in typeahead/search
This applies both to the typeahead in compose after you type a `:`,
and to the emoji search screen when you go to add a reaction to a
message.

The notable behavior change this gets us is nice behavior on spaces:
in your query you can now separate words with spaces, rather than
having to type underscores.  (It still works the same if you do type
an underscore.)

Fixes: #4636