mute: Request and consume new user_topic format in API
In this commit, we start requesting the new `user_topic` format for
the data describing the "muted topics" feature.

On new servers, that causes the new format to appear, and the old to
disappear.  Describe the API changes, and start consuming the new
format.

This should have no user-visible effect -- currently the new format
describes exactly the same information as the old.  Taking the
client plus (a properly behaving) server as one system, this change
is therefore NFC.

We don't call the change NFC, though, because it does change the
client's behavior as witnessed by the server.

(This change does switch from stream names to stream IDs in the
API... but only for data that we get from the server through the
event system, not for data we send or receive in any request outside
that system.  That means that (unless affected by some server bug)
it's always working with a stream name-to-ID mapping that exactly
matches what the server had at the time it sent that data.  So even
if a stream gets renamed concurrently with getting some data here,
there's no race and no bug.

OTOH when we *change* this at the server on the user's behalf, that
does suffer from such a race.  But that part of the API was fixed
long ago to accept stream IDs; we already have a TODO-server
comment, at `api.setTopicMute`, to start relying on that update.)

Fixes: #5380