users reducer tests [nfc]: Fill out user-update tests (still skipped)
The reducer still doesn't handle all these cases, so we can't
un-skip them yet. But now at least we're prepared to cover all the
cases once the reducer *does* handle them, soon.

See the list of cases in the doc:
  https://zulip.com/api/get-events#realm_user-update

The removed comment was wrong to say that we should exclude a test
for bot-owner changes. The users state doesn't include cross-realm
bots, it's true. But it does include regular bots, and those are the
bots that have owners.