api types: Let CrossRealmBot have profile_data, like User
Even if in fact a cross-realm bot never has a `profile_data` property,
this type with `profile_data?: …` is still accurate.  (These types
purely describe data we get from the server, not any expectations the
server has for any data it gets from us, so it's always safe to make
them less specific.)

And in general the CrossRealmBot vs. User distinction is more a
legacy wart than something that's doing us any good.  We've recently
made progress on reconciling them, so keep on in that direction.
Discussion:
  https://github.com/zulip/zulip-mobile/pull/5398#discussion_r890598309