type tests: Delete soon-redundant copies of react-nav type-tests
These exist because our existing type definitions for react-navigation
repeat most of their contents in the file for each package.  (Which is
a legacy from flow-typed, where for a long time it was believed to be
impossible for one libdef to import types from another.)

We're about to replace these definitions with versions generated from
the upstream libraries' own TS type definitions.  Those won't have
this redundancy, so this type we're testing will be exported only by
the one package that actually defines it.  Delete the tests that will
stop making sense.