tools/tsflower: Leave "[PATCH]" out of patches; keep "[tsflower]" etc.
By default, `git format-patch` adds prefixes like "[PATCH] " to the
subject line, and `git am` strips those prefixes as well as anything
else in brackets.  That means this "[tsflower] " prefix (meant to
classify this patch as one that could be obviated by a change in
TsFlower) gets lost on `tools/tsflower unpack && tools/tsflower pack`.

We could tell `git am` to strip only the PATCH prefixes, with
`--keep-non-patch`.  But better yet, tell `git format-patch` not to
add any such prefix, and then simply tell `git am` to take the
subject line as is.