nav types: Child navigators' navigation props have parent's methods too
For example, in HomeScreen we have a `navigation` prop that is a
MainTabsNavigationProp, because HomeScreen appears under the
navigator created in MainTabsScreen.  That prop has a method
`navigation.push` -- a stack-navigation method -- even though the
navigator is a tab navigator.  That's because the navigator's parent
is our main app navigator, a stack navigator.

Our type MainTabsNavigationProp hadn't reflected that, so the types
had said that `navigation.push` method didn't exist there.  Amend
that type, and similar ones, to reflect that it does.