android build: Move `mavenCentral()` a few lines down and exclude RN
Done to follow the template-app change in
facebook/react-native@046b02628, on the path to the RN v0.67
upgrade, with the fixup in facebook/react-native@2f8e52b52.

The fixup is aimed at projects that use RN's nightly builds. We
don't, but we might as well follow the template app. And using
nightly releases is mentioned in a (possibly early/outdated?) draft
for opting into RN's "new architecture":
  https://reactnative.dev/docs/next/new-architecture-app-intro#use-a-react-native-nightly-release
and I suppose one day we'll want to be using that new architecture.