flow-todo: Switch parser to @babel/eslint-parser, from babel-eslint
Oops, looks like we should have done this in e1833580a, for #4254.

I'm not sure how much we actually want to keep the `test:flow-todo`
command, but might as well bring it up to date in this way.

We're about to change it again, to hermes-eslint, along with our
other ESLint configs.