url: Fix some bugs in isUrlOnRealm
See the previous commit where we started pinning this down: we gave
this function a jsdoc, and commented about some false positives and
false negatives we identified.

Now, use our assumption that the `url` param is a valid URL string,
to use isUrlRelative, which requires that.

isUrlRelative is carefully defined to follow the spec, so
isUrlOnRealm is now much clearer that it does what it says on the
tin, modulo the remaining false-negative bug, which we'll fix next.