url [nfc]: Assume correct hrefs in messages; pin down isUrlOnRealm a bit
If isUrlOnRealm can assume its `url` param is a valid URL string,
then it'll be able to use the nicely defined function isUrlRelative,
which has that requirement. (We'll do that, coming up.)

Thankfully, it seems like its callers do pass well-formed URLs. They
operate on the `href` attribute of links and images in messages,
when one of those is pressed. And Anders says
  ( https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/Interpreting.20links.20in.20messages/near/1410864 ):

> The server should be responsible for sending you correct `href`
> attributes.

The new jsdoc seems to match the function's name, and the
requirements of its existing callsites:

- It's used to decide whether to send an authenticated request to
  get a "temporary file URL" (see src/api/tryGetFileTemporaryUrl)
  when the user clicks a link in a message.

- It's used to decide whether to include the auth credential when
  pointing react-native-photo-view's PhotoView component to an image
  on the Web, for the lightbox.

I've written down some false positives and false negatives in the
implementation, on the lines of code responsible for them. We'll fix
those next.